--- MANIFEST.in
@@ -1,11 +1,0 @@
-exclude .gitignore
-exclude .coverage
-exclude .travis.yml
-exclude .vscode
-include README.md
-exclude setup.cfg
-prune .cache
-prune .git
-prune build
-prune dist
-prune travis-scripts
