+++ teledump/exceptions/DumpingError.py
