+++ teledump/exceptions/__init__.py
