+++ teledump/filters/Filter.py
